leftmost skeleton tree (lst) routines
The leftmost skeleton tree (described in "Stronger Quickheaps"
(Navarro, Paredes, Poblete, and Sanders in International Journal
of Foundations of Computer Science, November 2011) is a structure
for implementing priority queues that has the efficiency of the
quickheap without being sensitive to input key distribution.

Included here is source for the LST routines and a test program
to exercise them, along with a simple makefile.