Add graphic showing the relative speed of binary heap and LST
The graphic shows the relative speed of binary heap and LST
for the three portions of the "cycle" test. In addition there
is an md file describing the sections of the test and the
operations performed in each, with some discussion of which
operations are to the advanage of one data structure or the
other.